+++ rebuild.sh
@@ -6,4 +6,5 @@ fi
 
 hg log --verbose | ./hgChangeLog.sh > changelog.md
 ./sbt clean test '+package' '+package-doc' '+package-src' '+publish-local'
+./maven-metadata.sh